refactor: project structure modify and rename files
- 폴더로 컴포넌트의 계층구조를 나눔
- 그리고 파일이름을 index.ts로 변경
