+++ src/component/Board/Tranding/index.css
