+++ src/component/Board/board.css
@@ -1,16 +1,16 @@
-[data-component='Board'] {
+[data-component="Board"] {
   margin-left: auto;
   margin-right: auto;
 }
 
 @media (max-width: 1919px) {
-  [data-component='Board'] {
+  [data-component="Board"] {
     width: 1376px;
   }
 }
 
 @media (max-width: 1440px) {
-  [data-component='Board'] {
+  [data-component="Board"] {
     width: 1024px;
   }
 }
