+++ src/component/Searching/SearchResult/index.ts
