+++ src/page/searchpage.ts
@@ -1,19 +1,18 @@
 import Component from "core/Component";
-import Searching from "component/Searching/searching";
+import Searching from "component/Searching";
 
-export default class SearchPage extends Component{
-  setup(){
+export default class SearchPage extends Component {
+  setup() {}
 
-  }
-  template(){
+  template() {
     return `
       <div data-component="Searching"></div>
-    `
+    `;
   }
-  mounted(){
-    const $searching = this.$target.querySelector('[data-component="Searching"]');
+
+  mounted() {
+    const $searching = this.$target.querySelector("[data-component='Searching']");
 
     new Searching($searching, {});
   }
-  
 }
